requirements.txt
datetime
google-cloud-logging
google-cloud-bigquery
google.cloud
google.cloud.logging
pandas==2.2.2
pandas-gbq==0.23.1
db-dtypes
google-cloud-secret-manager

